# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "NetworkSim API",
    "description": "Network Topology Simulator with Chaos Engineering capabilities.\n\n## Features\n- Create and manage network topologies\n- Deploy topologies to Kubernetes\n- Apply chaos engineering (delay, loss, partition, etc.)\n- Monitor network metrics and events\n- Run diagnostic tests",
    "contact": {
      "name": "NetworkSim Team"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "List events with filtering",
        "operationId": "list_events",
        "parameters": [
          {
            "name": "topology_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "post": {
        "tags": [
          "events"
        ],
        "summary": "Create a new event",
        "operationId": "create_event",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Event created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/events/stats": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get event statistics",
        "operationId": "event_stats",
        "parameters": [
          {
            "name": "topology_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event statistics",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
    "/api/topologies/{id}/events": {
      "get": {
        "tags": [
          "events"
        ],
        "summary": "Get events for a specific topology",
        "operationId": "list_topology_events",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Topology ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "topology_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "event_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "severity",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "source_type",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "until",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "nullable": true
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64",
              "nullable": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of topology events",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Topology not found"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateEventRequest": {
        "type": "object",
        "description": "Request body to create an event\nPayload to create a new event associated to an optional topology or source object.",
        "required": [
          "event_type",
          "title"
        ],
        "properties": {
          "description": {
            "type": "string",
            "description": "Optional detailed description",
            "example": "The link experienced continuous packet loss for 10s",
            "nullable": true
          },
          "event_subtype": {
            "type": "string",
            "description": "Optional subtype",
            "example": "packet-loss",
            "nullable": true
          },
          "event_type": {
            "type": "string",
            "description": "High-level event type (required)",
            "example": "link.down"
          },
          "metadata": {
            "type": "object",
            "description": "Optional JSON metadata object",
            "nullable": true
          },
          "severity": {
            "type": "string",
            "description": "Optional severity, defaults to `info`",
            "example": "warning",
            "nullable": true
          },
          "source_id": {
            "type": "string",
            "description": "Optional source identifier",
            "example": "link-1",
            "nullable": true
          },
          "source_type": {
            "type": "string",
            "description": "Optional source type (node, link, etc.)",
            "example": "link",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "Short title for the event (required)",
            "example": "Link down between A and B"
          },
          "topology_id": {
            "type": "string",
            "description": "Optional topology id to associate the event with",
            "example": "topology-1234",
            "nullable": true
          }
        }
      },
...
      "Event": {
        "type": "object",
        "description": "System event\nRecorded event occurring in the system or a topology. Contains human-readable title and optional structured metadata.",
        "required": [
          "id",
          "event_type",
          "severity",
          "title",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp of event creation (RFC3339)",
            "example": "2025-01-01T12:00:00Z"
          },
          "description": {
            "type": "string",
            "description": "Longer optional description",
            "example": "Detected sustained packet loss on interface eth0 to eth1",
            "nullable": true
          },
          "event_subtype": {
            "type": "string",
            "description": "Optional subtype for finer classification",
            "example": "packet-loss",
            "nullable": true
          },
          "event_type": {
            "type": "string",
            "description": "High-level event type (e.g. `link.down`, `node.reboot`)",
            "example": "link.down"
          },
          "id": {
            "type": "integer",
            "format": "int64",
            "description": "Database-assigned event identifier",
            "example": 1234
          },
          "metadata": {
            "type": "object",
            "description": "Optional structured metadata as JSON object",
            "nullable": true
          },
          "severity": {
            "type": "string",
            "description": "Severity as string (info|success|warning|error)",
            "example": "info"
          },
          "source_id": {
            "type": "string",
            "description": "Identifier of the source object",
            "example": "node-1",
            "nullable": true
          },
          "source_type": {
            "type": "string",
            "description": "Source object type (node, link, etc.)",
            "example": "link",
            "nullable": true
          },
          "title": {
            "type": "string",
            "description": "Short human-readable title",
            "example": "Link between A and B is down"
          },
          "topology_id": {
            "type": "string",
            "description": "Associated topology id if applicable",
            "example": "topology-1234",
            "nullable": true
          }
        }
      },
...
      "EventsResponse": {
        "type": "object",
        "description": "Events list response with pagination\nPaginated list of events matching the query",
        "required": [
          "events",
          "total",
          "has_more"
        ],
        "properties": {
          "events": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "description": "List of events"
          },
          "has_more": {
            "type": "boolean",
            "description": "Whether more pages are available"
          },
          "total": {
            "type": "integer",
            "format": "int64",
            "description": "Total number of matching events",
            "example": 42
          }
        }
      },
...
    }
  },
...
}